Check bounds on every indexed access (#14)
The "once OOB, stays OOB" behavior is problematic for the following
reasons:

- Implementations incur additional complexity to track an OOB state.

- TAs do not go into OOB state **unless** observed to be OOB via e.g.
  indexed access. This may be surprising. (The alternative of having
  each RAB/GSAB track all its views is equally distasteful due to
  implementation complexity.)

In contrast, doing a bounds check on every indexed access is simple to
explain. The tradeoff is that it makes indexed access slower, and that
performance can only recouped in an optimizing JIT that can hoist these
bounds checks.