Relax GSAB byte length accesses to unordered for get/set (#28)
Closes #27.

All other operations are SeqCst, while indexed get and set are relaxed
to Unordered. This aligns with wasm and also gives implementations
greater latitude in implementing indexed access on TAs backed by GSABs.

It is an invariant of that there is a single shared memory read of the
byteLength per buffer per higher level operation (e.g. "TA#slice"). To
make this easier across various AOs that do bounds checking, there is
the MakeIdempotentArrayBufferByteLength AO that returns an AC that makes
getting the byteLength idempotent.