Extend [Shared]ArrayBuffer instead of introducing new constructors (#37)
* Extend [Shared]ArrayBuffer instead of introducing new constructors

This is an attempt to square the circle to address feedback from
Moddable and other delegates' for adding new global constructors.

We can sidestep the "where should constructors live" discussion by
extending the behavior of existing ArrayBuffers.

Note that this is still opt-in: the behavior of the existing uses of
ArrayBuffers is *unchanged*. An call to the ArrayBuffer constructor with
a single argument results in a fixed-length buffer that can neither grow
nor shrink.

For implementations that wish to keep existing ArrayBuffer and
TypedArray code paths separate for e.g. ease of security audit,
resizable ArrayBuffers are still recommended to get different hidden
classes internally.