Normative: require internal slot in accessors
The existing `byteLength` accessor functions on both ArrayBuffer and
SharedArrayBuffer require the *this* value to have the
[[ArrayBufferData]] internal slot, as do the `maxByteLength` accessor
functions defined by this proposal.

Additionally, the abstract operations in use within the `resizable` and
`growable` accessor methods include invariants regarding the presence of
the [[ArrayBufferData]] internal slot.

Update the `resizable` and `growable` accessor methods to require the
[[ArrayBufferData]] internal slot to promote consistency with other APIs
and to satisfy the invariants of abstract operations.