Editorial: Remove outdated note (#47)
This note was added in commit 92c99420 to help explain the operation
that preceded it. Commit 08928ea1 subsequently re-ordered the invocation
of abstract operations in such a way as to obviate the note. (In the
current iteration of the algorithm, the [[ArrayBufferByteLength]]
internal slot is never read for detached ArrayBuffers.)