Normative: correct TypedArray boundary checks (#55)
The "byteOffsetEnd" variable is ostensibly intended to describe the
first invalid index in the underlying ArrayBuffer. When deriving this
value for fixed-length TypedArray instances, it should take into account
not only the instance's length and "element size" but also the
instance's offset into the ArrayBuffer.

When this value is equal to the underlying ArrayBuffer's "byte length,"
the TypedArray is within bounds.

Update the calculation of "byteOffsetEnd" for fixed-length TypedArray
instances, and label maximally-sized TypedArray instances as within
bounds.