Editorial: correct DataView length & boundary
Prior to this patch, the GetViewByteLength abstract operation returned
the length of the underlying buffer. This value would only be accurate
if the DataView had a byte offset of zero. Update the algorithm to
take the DataView's byte offset into account.

Prior to this patch, the IsViewOutOfBounds abstract operation relied on
the GetViewByteLength abstract operation to provide an accurate value
for the byte length of the DataView. The algorithm was incorrect when it
used the incorrect version of GetViewByteLength, but the correction to
GetViewByteLength is not sufficient to correct this abstract operation.
Update the algorithm to report whether the updated version of
GetViewByteLength returns a valid value.

Replace the reference to the non-existent specification variable "O"
with a reference to the specification variable "buffer".