Account for resize-in-the-middle for TA methods
Closes #73

At a high-level:

- Lengths are cached at the beginning of methods.
- Only in-bounds elements are copied.

Together, this means that the result TAs have the snapshotted length at
the beginning of the method. If the source TA's buffer shrunk due to
the species constructor, the elements between [newShrunkLen,
originalLen) in the result TA remain 0.

This is an alternative solution to #74. In #74, elements between
[newShrunkLen, originalLen) are set to the result of OOB [[Get]]s on the
source TA.  While doing so is maximally consistent with the Array
versions of these methods, it turns out to be really weird for
floating-point TAs. For OOB [[Get]]s on TAs returns `undefined`, and
setting `undefined` in Float32Arrays and Float64Arrays coerces to NaN.
This means extra complexity and code implementations have to do for an
edge case (resize-in-the-middle) that's bad practice anyway.