Editorial: simplify DataView validation
Previously, the variable named "byteLengthChecked" could contain a
specification label ("~empty~") or a Number value. Despite this, it was
only used to determine whether a single condition had been satisfied.
Using a complex type to describe a binary state made the algorithm
appear more complicated than it truly is.

Simplify the value in order to more clearly communicate the variable's
effect on the algorithm.