Editorial: simplify DataView validation
Previously, the variable named "byteLengthChecked" could contain a
specification label ("~empty~") or a Number value. Despite this, it was
only used to determine whether a single condition had been satisfied.
Using a complex type to describe a binary state made the algorithm
appear more complicated than it truly is.

In addition, that variable was defined solely to determine whether the
view's byte length needed to be validated. However, the conditions
determining the necessity of this validation can be inferred from other
specification variables.

Remove the "byteLengthChecked" variable and reference the value of
"viewByteLength" to determine when to perform validation. Although this
approach will encourage unnecessary validation (specifically: whenever a
byte length has not been specified for a non-resizable buffer), that
behavior is not observable and therefore implementations are free to
skip it.